Refactor needed types to @material-ui/types
